services/auth&auth/requirements.txt
Flask==2.2.2
authlib==1.2.0
requests==2.31.0
Werkzeug==2.2.2
redis==4.5.5
Flask_SQLAlchemy==3.0.3
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.6